Isolate function to update checkbox list
The function to update the checkbox list should be
isolated and added as a utility function in order
to use it in the test as just using the `updateIpaObject`
alone won't work to trigger the test when a given checkbox
is updated.
